Acquire images in leadup to recording
To account for the fact that the camera sometimes takes a bit to get
warmed up, this reordered the script so that the camera is collecting
the images all through the pre-recording phase, just not saving them.
Hopefully this will help address the artifacts when the camera first
starts up.