Initial commit of refactoring of Luke's photometry
In dealing with errors with Luke's photometry code, it was getting
difficult to work with given the heavy use of globals. I also wanted to
try to move away from using Gaussians to better improve the reliability
of the script, as they seem to be causing issues. A full rewrite seemed
more clear than trying to edits Luke's code infile.

Has about 30% of functionality currently. Primary magnitude function is
currently being worked on.