Implement temperature and humidity tracking
Also adds an entry in the gui details table that show the last reported
value of both the temperature and the humidity. Both are currently set
to log every 5 minutes.

A new log was created to serve as a dump for all this temperature and
humidity data, so that we can ensure that nothing fishy is going on and
reaching dangerous levels.