Add conditional output when temp fails
When the temperature sensor sends an error, the logging program
currently writes an NA to the log and the shared python file. This was
being read incorrectly in the gui display, and causing an error. Here we
explicitly check if the value is NA and adjust what is output
accordingly.