Add short clip writer class to utils
Adds a new class which is highly similar to the KeyClipWriter class but
which will not record any video over a certain duration. Additionally,
writing to disk is moved to take place entirely after the event has
occurred, so as to not be writing in cases when the event is already too
long.