Adapt to use ShortClipWriter instead
Rewrote portions to better utilize the ShortClipWriter so as to not save
needlessly long videos. Hopefully this will cut down on the amount of
false positives being saved. Also made some clean-up tweaks to the
bottom portion of the module to better let it be run as main, which now
works.