Fix potential memory overflow from long videos
Since the ShortClipWriter now writes everything upon the conclusion of
an event, all frame data is stored for the duration. In long cases such
as bugs wandering over the top of the dome, this may lead to long videos
whose queued frames fill up memory. This fix adjusts the main script to
check after each update if the toolong flag has been raised and end the
recording immediately at that point, freeing the queue.