Fix threaded videos saving properly
Worked out that much of the issue was from the reference to self.writer
within the threaded function that was causing frames to be written to
the next video if it happened quickly after another finished. Fixed that
as well as saving threads is a list to not have any conflicts there
either.